Move Plot component to components folder
