+++ src/components/Plot/Plot.js
@@ -1,15 +1,13 @@
-// Plot.js
-import React from 'react';
+import React, { Component } from 'react';
 import { XAxis, YAxis, CartesianGrid, LineChart, Tooltip, Line, ScatterChart, Scatter, Legend } from 'recharts';
 
-class Plot extends React.Component {
-    constructor() {
-        super();
-        this.state = {
-            // Color palette from Tableau Medium
-            colors: ["rgb(114,158,206)", "rgb(255,158,74)", "rgb(103,191,92)", "rgb(237,102,93)", "rgb(173,139,201)", "rgb(168,120,110)", "rgb(237,151,202)", "rgb(162,162,162)", "rgb(205,204,93)", "rgb(109,204,218)"],
-            options: {}
-        }
+import './Plot.css'
+
+class Plot extends Component {
+    state = {
+        // Color palette from Tableau Medium
+        colors: ["rgb(114,158,206)", "rgb(255,158,74)", "rgb(103,191,92)", "rgb(237,102,93)", "rgb(173,139,201)", "rgb(168,120,110)", "rgb(237,151,202)", "rgb(162,162,162)", "rgb(205,204,93)", "rgb(109,204,218)"],
+        options: {}
     }
 
     shouldComponentUpdate(nextProps, nextState) {
